+++ Tests/Integration/MSFT_WaitForNetworkTeam.config.ps1
@@ -1,0 +1,10 @@
+configuration MSFT_WaitForNetworkTeam_Config {
+    Import-DscResource -ModuleName NetworkingDsc
+    node localhost {
+        WaitForNetworkTeam Integration_Test {
+            Name             = $Node.Name
+            RetryIntervalSec = $Node.RetryIntervalSec
+            RetryCount       = $Node.RetryCount
+        }
+    }
+}