+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - MSFT_IPAddress:
   - Updated to allow setting multiple IP Addresses
     when one is already set - Fixes [Issue #323](https://github.com/PowerShell/NetworkingDsc/issues/323)
+- MSFT_WaitForNetworkTeam:
+  - Added a new resource to set the wait for a network team to become 'Up'.
 
 ## 5.7.0.0